+++ src/components/Avatar/AvatarGroup.js
@@ -1,3 +1,9 @@
+/**
+ * @license
+ * Copyright tuantvk. All Rights Reserved.
+ * Licensed under the MIT License. See License in the project root for license information.
+ */
+
 import React, { Component } from 'react';
 import {
   View,