+++ src/components/Badge/Badge.js
@@ -1,3 +1,9 @@
+/**
+ * @license
+ * Copyright tuantvk. All Rights Reserved.
+ * Licensed under the MIT License. See License in the project root for license information.
+ */
+
 import React, { Component } from 'react';
 import {
   View,
@@ -7,6 +13,7 @@ import {
   blue,
   white,
 } from '../../styles';
+import PropTypes from 'prop-types';
 
 class Badge extends Component {
   render() {
@@ -39,6 +46,13 @@ class Badge extends Component {
 Badge.defaultProps = {
   size: 24,
   color: white,
+}
+
+Badge.propTypes = {
+  size: PropTypes.number,
+  color: PropTypes.string,
+  bgColor: PropTypes.string,
+  value: PropTypes.string,
 }
 
 export default Badge;